requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.4.1
async-timeout==4.0.1
attrs==21.2.0
certifi==2021.5.30
cffi==1.14.6
charset-normalizer==2.0.6
cryptography==3.4.8
dateparser==1.1.0
defusedxml==0.7.1
Django==3.2.7
django-allauth==0.45.0
django-rest-auth==0.9.5
djangorestframework==3.12.4
djangorestframework-datatables==0.6.0
frozenlist==1.2.0
idna==3.2
iniconfig==1.1.1
multidict==5.2.0
oauthlib==3.1.1
packaging==21.3
pluggy==1.0.0
py==1.11.0
pycoingecko==2.2.0
pycparser==2.20
PyJWT==2.1.0
pyparsing==3.0.6
pytest==6.2.5
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
regex==2021.11.10
requests==2.26.0
requests-oauthlib==1.3.0
six==1.16.0
sqlparse==0.4.2
toml==0.10.2
typing_extensions==4.0.0
tzdata==2021.5
tzlocal==4.1
ujson==4.3.0
urllib3==1.26.7
websockets==9.1
yarl==1.7.2
